ci(workflows): Make Android, Native & Code-Analysis workflows trigger together
The goal is to make the Renovate pull requests to always trigger the 3 workflows
so the code coverage always validate the whole codebase.
Also increase the code coverage threshold to a value which allows the pipeline
to pass even if only 1 workflow was triggered.
Also rename the Docker build to contain only the base image name and the host
operating system.