fix(docker.sh): Fix open & close of file descriptor 3 in `pullDockerImage`
Summary:
Also improve the function `compileMobileRTInDockerContainer` by making it to
automatically detect current path in Windows host systems. This makes the
function to also works in Windows systems.
And change the container name used by `executeUnitTestsInDockerContainer` to
be different from the container used by `compileMobileRTInDockerContainer`. By
doing this change, it's possible to run a docker container to compile MobileRT
and another to execute the tests for the same image. And also made the function
`commitMobileRTDockerImage` to use the same docker container created by
`compileMobileRTInDockerContainer`, so the newly updated image continues to have
MobileRT pre-compiled with the latest version as well.

Test plan:
* Pipeline passes