fix(deps): update dependency com.android.tools.build:gradle to v8.3.1
Only use Gradle v8.3.1 when Android API is >= 21. And use v8.2.2 by default so
it's compatible with Android API 15.
Also update NDK version to v25.2.9519653 for the more recent versions of Android.
This update is useful because Github actions machines already bring this
version of NDK already installed out of the box:
https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2004-Readme.md#android