ci(android): Enable KVM before executing the Android tests in Ubuntu 22+
Summary:
By doing this, the Android emulator will be hardware accelerated on Ubuntu 22+
runners, which improves the performance by a lot.
Followed this guide:
https://github.com/marketplace/actions/android-emulator-runner#running-hardware-accelerated-emulators-on-linux-runners
This has been available since:
https://github.blog/changelog/2023-02-23-hardware-accelerated-android-virtualization-on-actions-windows-and-linux-larger-hosted-runners
Didn't enable KVM for Ubuntu 20.04 as well because the Android tests have been
failing consistenly on Android API 15 because of it.

Test plan:
* Pipeline passes