ci(android): Enable KVM before executing the Android tests in Ubuntu 20
Summary:
By doing this, the Android emulator will be hardware accelerated on Ubuntu 20
runners, which improves the performance by a lot.
Followed this guide:
https://github.com/marketplace/actions/android-emulator-runner#running-hardware-accelerated-emulators-on-linux-runners
This has been available since:
https://github.blog/changelog/2023-02-23-hardware-accelerated-android-virtualization-on-actions-windows-and-linux-larger-hosted-runners
Also increased the number of samples per pixel to be rendered on the tests of
class `CameraTest` because otherwise Ubuntu can be too fast rendering the small
scenes and make the tests fail validating the changes on the Rendering button
due to hardware acceleration.

Test plan:
* Pipeline passes