ci(android.yml): Add retry when executing the Android tests
Summary:
Add the parameter to select the maximum number of retries to execute the shell
command to the shell function `callCommandUntilSuccess`.
By doing this, it is now possible to use it in the github workflows to avoid
flakiness. It is now used in the execution of the Android tests.

Test plan:
* Pipeline passes