build(CMake): Remove usage of OpenMP completely
Summary:
OpenMP is no longer used by MobileRT because it was found out that for some
NDK versions OpenMP would make the Android app fail with weird errors when
mixing std::mutex with OpenMP threads.
Also replace usage of boost::mutex by std::mutex since now that it's known that
the issue of MobileRT on some NDK versions was the usage of mutex with OpenMP
threads, it's now possible to use std::mutex instead of boost.

Test plan:
* Pipeline passes