ci(workflows): Make the workflows only trigger if itself was modified
Summary:
Remove some workflows from being triggered when other workflows are changed.
This was done in order to avoid having multiple workflows being triggered
unnecessary. Before, some workflows like Android and Native were "tied" in
order to make sure the code coverage was always properly calculated because
it depends on Android, Native and Code Analysis workflows, but it seems a waste
of time since we can always manually trigger the workflows to calculate that
when necessary.

Test plan:
* Pipeline passes