build(gradle): Only build for `x86` CPU architecture by default
Summary
For the APKs, build them for `x86` and `armeabi-v7a` CPUs, so it works on most
devices.
Only build for `x86_64` CPUs when the Android API > 29, because all devices
have 64bit CPU when Android is that recent.
Also increase the number of threads used to build MobileRT on Code Analysis job.
And increase the parallelism when building MobileRT for Android builds, so the
jobs can be faster.
Finally, fix the update of Gradle version to v8.10 to be used by
Android API > 16.

Test plan:
* Pipeline passes