ci(android.yml): Change all jobs to debug
Summary
This change is necessary because the jobs in release mode started to fail
when the it was added the compiler flag `-DNDEBUG` to the `RelWithDebInfo` mode
which the Android builds use. Which means that by reducing the logs and enable
more optimizations made broke the release mode. It's necessary to investigate it.
Also added information to some logs.
And added the test name to the error message when a test fails, so it's more
obvious which test fail.
And changed some tests to use Regular Grid acceleration instead of BVH because
it seems that only the BVH is broken in release mode.
Finally fixed some linter issues on `scripts/helper_functions.sh` script.

Test plan:
* Pipeline passes