test(androidTest): Improve `UtilsContextT#waitUntil` by making it wait in milliseconds at a time
Summary
The `UtilsContextT#waitUntil` now uses `ViewActionWait#waitFor` which creates
a `ViewAction` to let the main thread (app) to continue processing some
milliseconds at a time, which is much more accurate that just sleeping for 1
second at a time like it was before. This change reduces the flakiness of the
tests for the cases where it starts and stops MobileRT very fast, like the
PreviewTest does.
Also change some tests to use BVH as it was before, because it's the current
fastest accelerator.

Test plan:
* Pipeline passes