test(androidTest): Improve Android tests to work on MacOS
Summary
Increase overall Android tests timeout so they can also be executed with MacOS
in Github actions since they are slower.
Also improve `UtilsContextT#waitUntil` by making it to wait for the Bitmap to
also be updated so it reduces flakiness on the tests that just start and stop
MobileRT immediately with a slower Android emulator.
And remove some logs that can make MobileRT too slow on slower Android emulators.

Test plan:
* Pipeline passes