fix(androidTest): Add `AbstractTest#testWatcher` to validate all closeActions when a test succeeds
Summary
Before this change, when a test failed, the next tests being executed would
always fail with timeout, even when the test timeout was increased to 10 minutes.
The issue happened because the test that failed would throw an exception
during `AbstractTest#tearDown` when asserting that the expected Intent was
indeed sent by MobileRT application (to mock the UI and allow to read some
OBJ files from internal and external SD storage).
This commit fixes the issue, and now when a test fails, the next tests are
executed as expected.

Test plan:
* Pipeline passes