test(androidTest): Catch exception thrown in MainRendererTest#createAndGetIndexOfShader
Summary
With this change all Android tests are still executed even if some test in
MainRendererTest fails with the error:
```
There was an error while creating the shader object: (0) 0x0
```
which means that the GLES v2 wasn't properly initialized.
Also remove some debug logs in `Renderer#renderScene` because it spams too much
on CI when executed in Debug mode.

Test plan:
* Pipeline passes