test(androidTest): Improve Intent assertion on Android tests
Summary
When a test is initialized, it now verifies if there are some Intents that need
to be validated, and if so, then it validates the Intents from previous tests
that are of actions ACTION_GET_CONTENT or ACTION_MAIN, so they don't interfere
in the current test validations.
Also improve the AbstractTest#tearDown by adding a timeout of 20 seconds when
waiting for the MainActivity to finish, to avoid having the test fail due to
test timeout because the MainActivity of test was stuck during the finish state.

Test plan:
* Pipeline passes