test(androidTest): Improve code & decrease spl to 1 in some tests
Summary
This commit improves some Android Instrumentation test code.
Also decreases the samples per light from 99 to 1 in the following
tests:
* PreviewTest#testPreviewScenePerspectiveCamera
* PreviewTest#testPreviewSceneOrthographicCamera
* UiTest#testUI
* UiTest#testClickRenderButtonManyTimesWithoutPreview
* UiTest#testClickRenderButtonManyTimesWithPreview
This change is to allow the Shader algorithm, specially the
Path Tracing, to render some change on the bitmap as fast as possible
during the tests, and thus can avoid some flakiness due to timeout.

Test plan:
* Pipeline passes