ci(android): Create Android Github workflow template
Summary
With this change the Android workflow now starts the Android Tests
and Unit Tests jobs immediately after the Build job finishes.
Before this commit, the Android Tests and Unit Tests only started
when all Build jobs finished first.
The goal of this change is to improve a bit the overall execution
time of the Android workflow.

Test plan:
* Pipeline passes