fix(mobilertapp): Make UtilsLogging#logException to only log exceptions
Summary:
This commit makes some small changes:
* Make UtilsLogging#logException to only log exceptions since Errors
are not recoverable.
* Remove unnecessary imports.
* Use ScheduledExecutorService#scheduleWithFixedDelay in
RenderTask#doInBackground so a new task is only created after the
previous one has finished. With this change, we avoid possible issues
when Android processes become cached (moved to background), since it
could be possible that tasks may unexpectedly execute hundreds or
thousands of times in quick succession when a process changes from
cached to uncached.

Test plan:
* Pipeline passes